Feedback on how to go forward
This will serve as a basis for moving forward. I have delegated tasks that need to be completed with ❗ markers. This will make it easier to identify where and what needs to be finished. Whoever is assigned to a ❗ task should specify that they are working on it to prevent multiple people from working on the same task.